Add files via upload
College Database
A solo project that is a database created with SQL containing a college’s info. It includes
Students names, IDs, courses and grades. It also includes Departments; their names and
IDs. Also Instructors qualifications, IDs and the departments they belong too. The
courses and the topics related to them that the students study are also included.