Add files via upload
A solo project that is a dashboard created with Microsoft Excel that represents bike sales
according to dif erent metrics. This project was an important task that was required as a
part of the Data Analysis Course I took with the Ministry of Communications. I applied
all of what I learned on the topic of doing Data Analysis with Microsoft Excel.